+++ src/routes/+layout.svelte
@@ -3,7 +3,7 @@
 	import { Toaster } from '$lib/components/ui/sonner';
 	import { ModeWatcher } from 'mode-watcher';
 	import '../app.css';
-	import Header from './Header.svelte';
+	import Header from './_components/Header.svelte';
 
 	let { children } = $props();
 </script>
