+++ src/routes/+page.svelte
@@ -2,8 +2,8 @@
 	import userGuideEN from '$lib/assets/User-Guide-en.pdf';
 	import userGuideFR from '$lib/assets/User-Guide-fr.pdf';
 	import type { PageProps } from './$types';
-	import CreateProject from './CreateProject.svelte';
-	import ProjectsList from './ProjectsList.svelte';
+	import CreateProject from './_components/CreateProject.svelte';
+	import ProjectsList from './_components/ProjectsList.svelte';
 
 	let { data }: PageProps = $props();
 </script>
