+++ src/routes/_components/Header.svelte
@@ -5,7 +5,7 @@
 	import MoonIcon from '@lucide/svelte/icons/moon';
 	import SunIcon from '@lucide/svelte/icons/sun';
 	import { toggleMode } from 'mode-watcher';
-	import HeaderRegionsDropdown from './projects/[project]/regions/[region]/HeaderRegionsDropdown.svelte';
+	import HeaderRegionsDropdown from '../projects/[project]/regions/[region]/_components/HeaderRegionsDropdown.svelte';
 </script>
 
 <header>
